Added a zip task to build.gradle (#68)
* Changed TestAppBuilder to produce a zip file in addition to an apk

* Separated zipping code into a ZipUtils class

* Fixed missing script reference in test scene

* Moved test app zipping to gradle file

* Added buildTestApp and zipTestApp to automated tests

* Removed the option to specify the output file path
