+++ GooglePlayInstant/Editor/PlayInstantPublisher.cs
@@ -49,20 +49,7 @@ namespace GooglePlayInstant.Editor
             }
 
             // Zip creation is fast enough so call jar synchronously rather than wait for post build AppDomain reset.
-            var arguments = string.Format(
-                "cvf {0} -C {1} {2}",
-                CommandLine.QuotePathIfNecessary(zipFilePath),
-                CommandLine.QuotePathIfNecessary(baseApkDirectory),
-                BaseApkFileName);
-            var result = CommandLine.Run(JavaUtilities.JarBinaryPath, arguments);
-            if (result.exitCode == 0)
-            {
-                Debug.LogFormat("Created ZIP containing base.apk: {0}", zipFilePath);
-            }
-            else
-            {
-                PlayInstantBuilder.LogError(string.Format("Zip creation failed: {0}", result.message));
-            }
+            ZipUtils.CreateZipFile(baseApkDirectory, BaseApkFileName, zipFilePath);
         }
     }
 }
