+++ GooglePlayInstant/Samples/TestApp/Scenes/TestScene.unity
@@ -965,7 +965,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1647103411}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e3c23ec4faf7b464da2e57bcccd96417, type: 3}
+  m_Script: {fileID: 11500000, guid: 397d5161b5aa84e568b0e91b35d9ea84, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!4 &1647103413
