+++ build.gradle
@@ -104,16 +104,20 @@ task exportSampleSphereBlast {
     }
 }
 
+task zipTestApp(type: Zip) {
+   mustRunAfter 'buildTestApp'
+   from playInstantBuildPath
+   include "testapp.apk"
+   archiveName "testapp.zip"
+   destinationDir file(playInstantBuildPath)
+}
 
 task buildTestApp {
     description 'Creates the instant app test project, then builds a test app'
     def buildFunction = 'GooglePlayInstant.Samples.TestApp.Editor.TestAppBuilder.Build'
     def buildPath = "${playInstantBuildPath}/testapp"
     def logFile = file("testapp.log").getPath()
-    
-    //Specify output file path relative to the root path using -PoutputFile
-    def outputFileArg = project.hasProperty('outputFile') ? outputFile : "${playInstantBuildPath}/testapp.apk"
-    outputFileArg = file(outputFileArg)
+    def outputFile = file("${playInstantBuildPath}/testapp.apk");
         
     doFirst {
         copyFiles(buildPath, ['Tests'])
@@ -123,7 +127,7 @@ task buildTestApp {
         }
     }
     doLast {
-        println "Building to ${outputFileArg}"
+        println "Building to ${outputFile}"
         def argv = [
             '-batchmode',
             '-noGraphics',
@@ -134,7 +138,7 @@ task buildTestApp {
             '-buildTarget',
             'Android',
             '-outputFile',
-            outputFileArg,
+            outputFile,
             '-executeMethod',
             buildFunction,
             '-quit'
@@ -182,6 +186,8 @@ task clean(type: Delete) {
 task automatedTests {
     dependsOn 'exportPlugin'
     dependsOn 'tests'
+    dependsOn 'buildTestApp'
+    dependsOn 'zipTestApp'
 }
 
 project.defaultTasks = ['tests', 'exportPlugin']
